Fix Mono function/extension split
This commit renames MonoWhenFunctions to MonoFunctions since it also
contains a zip function. This should not break compatibility since
imports are using only package and function name. It also removes some
inconsistent and duplicated tests.
